+++ Sources/SuperwallKit/Models/Paywall/Paywall.swift
@@ -64,6 +64,10 @@ struct Paywall: Codable {
       products = Self.makeProducts(from: productItems)
     }
   }
+
+  /// Indicates whether scrolling is enabled on the webview.
+  var isScrollEnabled: Bool
+
   // MARK: - Added by client
 
   var responseLoadingInfo: LoadingInfo
@@ -140,6 +144,7 @@ struct Paywall: Codable {
     case computedPropertyRequests = "computedProperties"
     case surveys
     case manifest
+    case isScrollEnabled
 
     case responseLoadStartTime
     case responseLoadCompleteTime
@@ -256,6 +261,7 @@ struct Paywall: Codable {
     computedPropertyRequests = throwableComputedPropertyRequests.compactMap { try? $0.result.get() }
 
     manifest = try values.decodeIfPresent(ArchiveManifest.self, forKey: .manifest)
+    isScrollEnabled = try values.decodeIfPresent(Bool.self, forKey: .isScrollEnabled) ?? true
   }
 
   func encode(to encoder: Encoder) throws {
@@ -312,6 +318,7 @@ struct Paywall: Codable {
     }
 
     try container.encodeIfPresent(manifest, forKey: .manifest)
+    try container.encodeIfPresent(isScrollEnabled, forKey: .isScrollEnabled)
   }
 
 
@@ -368,7 +375,8 @@ struct Paywall: Codable {
     localNotifications: [LocalNotification] = [],
     computedPropertyRequests: [ComputedPropertyRequest] = [],
     surveys: [Survey] = [],
-    manifest: ArchiveManifest? = nil
+    manifest: ArchiveManifest? = nil,
+    isScrollEnabled: Bool
   ) {
     self.databaseId = databaseId
     self.identifier = identifier
@@ -401,6 +409,7 @@ struct Paywall: Codable {
     self.surveys = surveys
     self.products = Self.makeProducts(from: productItems)
     self.manifest = manifest
+    self.isScrollEnabled = isScrollEnabled
   }
 
   func getInfo(fromEvent: EventData?) -> PaywallInfo {
@@ -435,7 +444,8 @@ struct Paywall: Codable {
       localNotifications: localNotifications,
       computedPropertyRequests: computedPropertyRequests,
       surveys: surveys,
-      presentation: presentation
+      presentation: presentation,
+      isScrollEnabled: isScrollEnabled
     )
   }
 
@@ -486,7 +496,8 @@ extension Paywall: Stubbable {
       webviewLoadingInfo: .init(),
       productsLoadingInfo: .init(),
       shimmerLoadingInfo: .init(),
-      paywalljsVersion: ""
+      paywalljsVersion: "",
+      isScrollEnabled: true
     )
   }
 }